+++ lib/boruta/oauth/authorization/client.ex
@@ -3,6 +3,14 @@ defmodule Boruta.Oauth.Authorization.Client do
   Check against given params and return the corresponding client
   """
 
+  defmodule Token do
+    @moduledoc false
+
+    use Joken.Config
+
+    def token_config, do: %{}
+  end
+
   alias Boruta.ClientsAdapter
   alias Boruta.Oauth.Client
   alias Boruta.Oauth.Error
@@ -15,16 +23,16 @@ defmodule Boruta.Oauth.Authorization.Client do
       {:ok, %Boruta.Oauth.Client{...}}
   """
   @spec authorize(
-          [id: String.t(), secret: String.t(), grant_type: String.t()]
+          [id: String.t(), source: map(), grant_type: String.t()]
           | [
               id: String.t(),
-              secret: String.t() | nil,
+              source: map() | nil,
               redirect_uri: String.t(),
               grant_type: String.t()
             ]
           | [
               id: String.t(),
-              secret: String.t() | nil,
+              source: map() | nil,
               redirect_uri: String.t(),
               grant_type: String.t(),
               code_verifier: String.t()
@@ -39,11 +47,11 @@ defmodule Boruta.Oauth.Authorization.Client do
                :redirect_uri => nil,
                :status => :unauthorized
              }}
-  def authorize(id: id, secret: secret, grant_type: grant_type)
+  def authorize(id: id, source: source, grant_type: grant_type)
       when not is_nil(id) do
     with %Client{} = client <- ClientsAdapter.get_client(id),
          true <- Client.grant_type_supported?(client, grant_type),
-         {:ok, client} <- maybe_check_client_secret(client, secret, grant_type) do
+         {:ok, client} <- maybe_check_client_secret(client, source, grant_type) do
       {:ok, client}
     else
       false ->
@@ -54,22 +62,30 @@ defmodule Boruta.Oauth.Authorization.Client do
            error_description: "Client do not support given grant type."
          }}
 
-      _ ->
+      nil ->
         {:error,
          %Error{
            status: :unauthorized,
            error: :invalid_client,
            error_description: "Invalid client_id or client_secret."
          }}
+
+      {:error, reason} ->
+        {:error,
+         %Error{
+           status: :unauthorized,
+           error: :invalid_client,
+           error_description: reason
+         }}
     end
   end
 
-  def authorize(id: id, secret: secret, redirect_uri: redirect_uri, grant_type: grant_type)
+  def authorize(id: id, source: source, redirect_uri: redirect_uri, grant_type: grant_type)
       when not is_nil(id) and not is_nil(redirect_uri) do
     with %Client{} = client <- ClientsAdapter.get_client(id),
          :ok <- Client.check_redirect_uri(client, redirect_uri),
          true <- Client.grant_type_supported?(client, grant_type),
-         {:ok, client} <- maybe_check_client_secret(client, secret, grant_type) do
+         {:ok, client} <- maybe_check_client_secret(client, source, grant_type) do
       {:ok, client}
     else
       false ->
@@ -92,7 +108,7 @@ defmodule Boruta.Oauth.Authorization.Client do
 
   def authorize(
         id: id,
-        secret: secret,
+        source: source,
         redirect_uri: redirect_uri,
         grant_type: grant_type,
         code_verifier: code_verifier
@@ -102,7 +118,7 @@ defmodule Boruta.Oauth.Authorization.Client do
          :ok <- Client.check_redirect_uri(client, redirect_uri),
          :ok <- validate_pkce(client, code_verifier),
          true <- Client.grant_type_supported?(client, grant_type),
-         {:ok, client} <- maybe_check_client_secret(client, secret, grant_type) do
+         {:ok, client} <- maybe_check_client_secret(client, source, grant_type) do
       {:ok, client}
     else
       false ->
@@ -140,20 +156,123 @@ defmodule Boruta.Oauth.Authorization.Client do
      }}
   end
 
-  defp maybe_check_client_secret(client, secret, grant_type) do
+  defp maybe_check_client_secret(client, source, grant_type) do
     case Client.should_check_secret?(client, grant_type) do
       false ->
         {:ok, client}
 
       true ->
-        case Client.check_secret(client, secret) do
-          :ok ->
-            {:ok, client}
+        with {:ok, secret} <- extract_secret(source, client) do
+          case Client.check_secret(client, secret) do
+            :ok ->
+              {:ok, client}
 
-          _ ->
-            {:error, "Invalid client_secret."}
+            {:error, _error} ->
+              {:error, "Invalid client_id or client_secret."}
+          end
         end
     end
+  end
+
+  defp extract_secret(source, client), do: do_extract_secret(source, client, nil)
+
+  defp do_extract_secret(_source, %Client{token_endpoint_auth_methods: []}, nil),
+    do: {:error, "No client authentication method found for given client."}
+
+  defp do_extract_secret(_source, %Client{token_endpoint_auth_methods: []}, message),
+    do: {:error, message}
+
+  defp do_extract_secret(
+         source,
+         %Client{token_endpoint_auth_methods: ["client_secret_basic" | methods]} = client,
+         _message
+       ) do
+    case source[:type] do
+      "basic" ->
+        {:ok, source[:value]}
+
+      _ ->
+        message = "Given client expects the credentials to be provided with BasicAuth."
+        do_extract_secret(source, %{client | token_endpoint_auth_methods: methods}, message)
+    end
+  end
+
+  defp do_extract_secret(
+         source,
+         %Client{token_endpoint_auth_methods: ["client_secret_post" | methods]} = client,
+         _message
+       ) do
+    case source[:type] do
+      "post" ->
+        {:ok, source[:value]}
+
+      _ ->
+        message = "Given client expects the credentials to be provided with POST body parameters."
+        do_extract_secret(source, %{client | token_endpoint_auth_methods: methods}, message)
+    end
+  end
+
+  defp do_extract_secret(
+         source,
+         %Client{
+           secret: secret,
+           token_endpoint_auth_methods: ["client_secret_jwt" | methods],
+           token_endpoint_jwt_auth_alg: alg
+         } = client,
+         _message
+       )
+       when alg in ["HS256", "HS364", "HS512"] and is_binary(secret) do
+    signer = Joken.Signer.create(alg, secret)
+
+    case {source[:type], Token.verify(source[:value], signer)} do
+      {"jwt", {:ok, _claims}} ->
+        {:ok, secret}
+
+      {"jwt", {:error, _error}} ->
+        message = "The given client secret jwt does not match signature key."
+
+        do_extract_secret(source, %{client | token_endpoint_auth_methods: methods}, message)
+
+      {_, _} ->
+        message = "Given client expects the credentials to be provided with a jwt assertion."
+        do_extract_secret(source, %{client | token_endpoint_auth_methods: methods}, message)
+    end
+  end
+
+  defp do_extract_secret(
+         source,
+         %Client{
+           jwt_public_key: jwt_public_key,
+           secret: secret,
+           token_endpoint_auth_methods: ["private_key_jwt" | methods],
+           token_endpoint_jwt_auth_alg: alg
+         } = client,
+         _message
+       )
+       when alg in ["RS256", "RS364", "RS512"] and is_binary(jwt_public_key) do
+    signer = Joken.Signer.create(alg, %{"pem" => jwt_public_key})
+
+    case {source[:type], Token.verify(source[:value], signer)} do
+      {"jwt", {:ok, _claims}} ->
+        {:ok, secret}
+
+      {"jwt", {:error, _error}} ->
+        message = "The given client secret jwt does not match signature key."
+
+        do_extract_secret(source, %{client | token_endpoint_auth_methods: methods}, message)
+
+      {_, _} ->
+        message = "Given client expects the credentials to be provided with a jwt assertion."
+        do_extract_secret(source, %{client | token_endpoint_auth_methods: methods}, message)
+    end
+  end
+
+  defp do_extract_secret(source, client, _) do
+    do_extract_secret(
+      source,
+      %{client | token_endpoint_auth_methods: []},
+      "Bad client jwt authentication method configuration (jwks and token endpoint jwt auth algorithm do not match)."
+    )
   end
 
   defp validate_pkce(%Client{pkce: false}, _code_verifier), do: :ok