+++ lib/boruta/oauth/requests/refresh_token_resquest.ex
@@ -8,14 +8,17 @@ defmodule Boruta.Oauth.RefreshTokenRequest do
   """
   @type t :: %__MODULE__{
           client_id: String.t(),
-          client_secret: String.t(),
+          client_authentication: %{
+            type: String.t(),
+            value: String.t()
+          },
           refresh_token: String.t(),
           scope: String.t(),
           grant_type: String.t()
         }
-  @enforce_keys [:client_id, :client_secret, :refresh_token]
+  @enforce_keys [:client_id, :client_authentication, :refresh_token]
   defstruct client_id: nil,
-            client_secret: nil,
+            client_authentication: nil,
             refresh_token: nil,
             scope: "",
             grant_type: "refresh_token"