+++ priv/repo/migrations/20221108102337_optional_public_key_for_oauth_clients.exs
@@ -1,0 +1,15 @@
+defmodule Boruta.Repo.Migrations.OptionalPublicKeyForOauthClients do
+  use Ecto.Migration
+
+  def up do
+    alter table(:oauth_clients) do
+      modify :public_key, :text, null: true
+    end
+  end
+
+  def down do
+    alter table(:oauth_clients) do
+      modify :public_key, :text, null: false
+    end
+  end
+end