+++ priv/repo/migrations/20221118155336_add_token_endpoint_auth_method_to_oauth_clients.exs
@@ -1,0 +1,10 @@
+defmodule Boruta.Repo.Migrations.AddTokenEndpointAuthMethodToOauthClients do
+  use Ecto.Migration
+
+  def change do
+    alter table(:oauth_clients) do
+      add :token_endpoint_auth_methods, {:array, :string}, null: false,
+        default: ["client_secret_basic", "client_secret_post"]
+    end
+  end
+end