+++ test/mix/tasks/boruta.gen.controllers_test.exs
@@ -5,6 +5,7 @@ defmodule Mix.Tasks.Boruta.Gen.ControllersTest do
     "controllers/oauth/authorize_controller.ex",
     "controllers/openid/authorize_controller.ex",
     "controllers/openid/jwks_controller.ex",
+    "controllers/openid/userinfo_controller.ex",
     "controllers/oauth/token_controller.ex",
     "controllers/oauth/introspect_controller.ex",
     "controllers/oauth/revoke_controller.ex",